% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.r
\name{CorrMatrixNegFixed}
\alias{CorrMatrixNegFixed}
\title{Simulation Functions}
\usage{
CorrMatrixNegFixed(blocks, a, corr, off)
}
\arguments{
\item{blocks}{blocks}

\item{a}{a}

\item{corr}{corr}

\item{off}{off}
}
\description{
Simulation Functions
}
