% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sim}
\alias{data_sim}
\title{Simulated data}
\format{
An object of class \code{data.frame} with 160 rows and 77 columns.
}
\usage{
data_sim
}
\description{
A dataset containing simulated data
}
\keyword{datasets}
