% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{ErrorsComputation}
\alias{ErrorsComputation}
\title{Compute the Errors for the Different Imputation Methods}
\usage{
ErrorsComputation(trunc, corr, euc, miss, complete)
}
\arguments{
\item{trunc}{truncation}

\item{corr}{corr}

\item{euc}{euc}

\item{miss}{missing}

\item{complete}{complete}
}
\description{
Compute the Errors for the Different Imputation Methods
}
