% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{EstimatesComputation}
\alias{EstimatesComputation}
\title{Based on the MLE Functions (mklhood) and NewtonRaphson Function}
\usage{
EstimatesComputation(missingdata, perc, iter = 50)
}
\arguments{
\item{missingdata}{matrix}

\item{perc}{perc}

\item{iter}{iterations in NR algorithm}
}
\description{
(NewtonRaphsonLike), This function estimates the MEAN and SD from the Truncated using Newton Raphson.
}
