% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{glmnet_pred}
\alias{glmnet_pred}
\title{glmnet prediction}
\usage{
glmnet_pred(x, y, alpha = 0.5, lambda = 0.01)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{alpha}{alpha}

\item{lambda}{lambda}
}
\description{
glmnet prediction
}
