% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{imputeKNN}
\alias{imputeKNN}
\title{imputeKNN}
\usage{
imputeKNN(
  data,
  k,
  distance = "correlation",
  rm.na = TRUE,
  rm.nan = TRUE,
  rm.inf = TRUE,
  perc = 1,
  ...
)
}
\arguments{
\item{data}{data}

\item{k}{k}

\item{distance}{distance}

\item{rm.na}{remove NA}

\item{rm.nan}{remove NaN}

\item{rm.inf}{remove Inf}

\item{perc}{percentage}

\item{...}{additional options}
}
\description{
This Function imputes the data based on KNN-CORRELATION or
KNN-TRUNCATION. The Parameter Estimates based on the Truncated
Normal from EstimateComputation function is run on this function
}
