% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{KNNEuc}
\alias{KNNEuc}
\title{This Function imputes the data BASED on KNN-EUCLIDEAN}
\usage{
KNNEuc(data, k, rm.na = TRUE, rm.nan = TRUE, rm.inf = TRUE)
}
\arguments{
\item{data}{data set to be imputed}

\item{k}{number of neighbors for imputing values}

\item{rm.na}{remove NA}

\item{rm.nan}{remove NaN}

\item{rm.inf}{remove Inf}
}
\description{
This Function imputes the data BASED on KNN-EUCLIDEAN
}
