% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{lm_pred}
\alias{lm_pred}
\title{Prediction functions}
\usage{
lm_pred(x, y)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\description{
Collection of prediction functions for different models
}
\details{
lm prediction
}
