% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{MAR_generate}
\alias{MAR_generate}
\title{Miscellaneous functions}
\usage{
MAR_generate(data, mis_prop = 0.5)
}
\arguments{
\item{data}{data}

\item{mis_prop}{missing proportion}
}
\description{
Collection of miscellaneous functions
}
\details{
Missing at Random Function
}
