% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp.R
\name{miss_init}
\alias{miss_init}
\title{Initialise the missing data}
\usage{
miss_init(miss_data, method = c("lsym", "qrilc", "rsym")[1])
}
\arguments{
\item{miss_data}{missing data}

\item{method}{method of initialisation of the missing data}
}
\description{
lsym will draw samples from the right tail of the distribution and transformed to the left tail
}
