% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.r
\name{MissingData}
\alias{MissingData}
\title{Create MissingValues Datasets}
\usage{
MissingData(data, totalmiss, mar, perc)
}
\arguments{
\item{data}{data}

\item{totalmiss}{total missing}

\item{mar}{mar}

\item{perc}{percentage}
}
\description{
Create MissingValues Datasets
}
