% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{mklhood}
\alias{mklhood}
\title{MLE for the trunckated normal distribution}
\usage{
mklhood(data, t, ...)
}
\arguments{
\item{data}{numeric vector}

\item{t}{truncation limits}

\item{...}{additional options}
}
\description{
Creating a Function that Returns the Log Likelihood, Gradient and Hessian Functions
}
