% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{MNAR_gen_imp}
\alias{MNAR_gen_imp}
\title{Imputation performance evaluation}
\usage{
MNAR_gen_imp(
  data_c,
  mis_var_prop = seq(0.1, 0.8, 0.1),
  var_mis_prop = seq(0.1, 0.8, 0.1),
  impute_list = c("kNN_wrapper", "SVD_wrapper", "HM_wrapper", "QRILC_wrapper"),
  cores = 5
)
}
\arguments{
\item{data_c}{data}

\item{mis_var_prop}{missing variable proportion}

\item{var_mis_prop}{variable missing proportion}

\item{impute_list}{list of imputation}

\item{cores}{number of cores}
}
\description{
Collection of functions for Imputation performance evaluation
}
\details{
MNAR generation and imputation
}
