% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{MNAR_generate}
\alias{MNAR_generate}
\title{MNAR imputation compare}
\usage{
MNAR_generate(data_c, mis_var = 0.5, var_prop = seq(0.3, 0.6, 0.1))
}
\arguments{
\item{data_c}{data to compare}

\item{mis_var}{missing variables}

\item{var_prop}{proportion of variables}
}
\description{
MNAR imputation compare
}
