% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot 4 ggplot2}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{additional options}

\item{plotlist}{list of plots}

\item{file}{file name}

\item{cols}{number of columns}

\item{layout}{layout}
}
\description{
Multiplot 4 ggplot2
}
