% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDim4jecs.R
\docType{package}
\name{NDim4jecs}
\alias{NDim4jecs}
\alias{NDim4jecs-package}
\title{NDim4jecs}
\description{
This is a package to implement GSimp codes. GSimp is created by Wei et al. and made as a R package for the use in the Japan Environment and Children's Study.
}
\author{
\strong{Maintainer}: Shoji Nakayama \email{fabre@nies.go.jp}

}
