% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{NewtonRaphsonLike}
\alias{NewtonRaphsonLike}
\title{Newton Raphson Function}
\usage{
NewtonRaphsonLike(lhood, p, tol = 1e-07, maxit = 100)
}
\arguments{
\item{lhood}{lhood}

\item{p}{p}

\item{tol}{tolerance}

\item{maxit}{max iterations}
}
\description{
This takes in the Objects Returned from mklhood Function above
}
