% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{nrmse_col}
\alias{nrmse_col}
\title{column-wise NRMSE calculation}
\usage{
nrmse_col(data_imp, data_miss, data_true)
}
\arguments{
\item{data_imp}{data imputation}

\item{data_miss}{data missing}

\item{data_true}{true data}
}
\description{
column-wise NRMSE calculation
}
