% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{NRMSE_list}
\alias{NRMSE_list}
\title{Calculate NRMSE on a list of imputed data}
\usage{
NRMSE_list(imp_list, miss_df, method_names)
}
\arguments{
\item{imp_list}{imputation list}

\item{miss_df}{missing data df}

\item{method_names}{names of methods}
}
\description{
Calculate NRMSE on a list of imputed data
}
