% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{NRMSE_rank_cal_plot}
\alias{NRMSE_rank_cal_plot}
\title{NRMSE rank calculate and plot}
\usage{
NRMSE_rank_cal_plot(
  impute_results,
  plot = T,
  x = "Miss_Prop",
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{results of imputation}

\item{plot}{boolean indicator if plot is done or not}

\item{x}{label of x-axis}

\item{colors}{colours}

\item{shapes}{shapes}
}
\description{
NRMSE rank calculate and plot
}
