% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{PCA_pro_list}
\alias{PCA_pro_list}
\title{Calculate PCA_Pro on a list of imputed data}
\usage{
PCA_pro_list(imp_list, nPCs = 2, method_names)
}
\arguments{
\item{imp_list}{imputation list}

\item{nPCs}{number of PCs}

\item{method_names}{names of methods}
}
\description{
Calculate PCA_Pro on a list of imputed data
}
