% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{pre_processing_GS_wrapper}
\alias{pre_processing_GS_wrapper}
\title{pre_processing_GS_wrapper}
\usage{
pre_processing_GS_wrapper(data)
}
\arguments{
\item{data}{data}
}
\description{
pre_processing_GS_wrapper
}
