% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{Procrustes_cal_plot}
\alias{Procrustes_cal_plot}
\title{Procrustes and plot}
\usage{
Procrustes_cal_plot(
  impute_results,
  DR = "PCA",
  lg = F,
  nPCs = 2,
  outcome = NULL,
  x = "Miss_Prop",
  plot = T,
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{imputation results}

\item{DR}{DR}

\item{lg}{lg}

\item{nPCs}{number of principal compornents}

\item{outcome}{outcome}

\item{x}{x-axis label}

\item{plot}{plot or not}

\item{colors}{colours}

\item{shapes}{shapes}
}
\description{
Procrustes and plot
}
