% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_data}
\alias{real_data}
\title{Real data}
\format{
An object of class \code{data.frame} with 198 rows and 77 columns.
}
\usage{
real_data
}
\description{
A dataset containing real data
}
\keyword{datasets}
