% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{rf_pred}
\alias{rf_pred}
\title{random forest prediction}
\usage{
rf_pred(x, y, ntree = 200, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{ntree}{number of trees}

\item{...}{additional options}
}
\description{
random forest prediction
}
