% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{RF_wrapper}
\alias{RF_wrapper}
\title{Imputation wrappers}
\usage{
RF_wrapper(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{additional options}
}
\description{
Collection of imputation wrappers
}
\details{
Random forest wrapper
}
