% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imput_funcs.r
\name{Rmse}
\alias{Rmse}
\title{Root Mean Squared Error Function}
\usage{
Rmse(imp, mis, true, norm = FALSE)
}
\arguments{
\item{imp}{imputation}

\item{mis}{missing}

\item{true}{true data}

\item{norm}{normal}
}
\description{
Root Mean Squared Error Function
}
