% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{Gibbs sampling functions}
\usage{
rnorm_trunc(n, mu, std, lo = -Inf, hi = Inf)
}
\arguments{
\item{n}{Number of samples}

\item{mu}{mean for the normal distribution}

\item{std}{standard deviation fro the normal distribution}

\item{lo}{lower end of the truncated distribution}

\item{hi}{upper end of the truncated distribution}
}
\description{
Collection of functions for Gibbs sampling functions
}
\details{
Draw n samples from a truncated normal distribution \code{N(mu, std^2|[lo, hi])} ##
}
