% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{scale_recover}
\alias{scale_recover}
\title{Scale and recover}
\usage{
scale_recover(data, method = "scale", param_df = NULL)
}
\arguments{
\item{data}{data}

\item{method}{method}

\item{param_df}{parameter df}
}
\description{
Scale and recover
}
