% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{sim_QRILC_wrapper}
\alias{sim_QRILC_wrapper}
\title{sim QRILC wrapper}
\usage{
sim_QRILC_wrapper(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{additional options}
}
\description{
sim QRILC wrapper
}
