% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.r
\name{SimulatedData}
\alias{SimulatedData}
\title{Simulated Data}
\usage{
SimulatedData(n, p, covar, low, high)
}
\arguments{
\item{n}{size}

\item{p}{metabolites}

\item{covar}{covariates}

\item{low}{low}

\item{high}{high}
}
\description{
This function Simulates complete data sets.
We are creating a dataset of size N (samples) by p (Metabolites)
}
