% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp.R
\name{single_impute_iters}
\alias{single_impute_iters}
\title{Single missing variable imputation based on Gibbs sampler}
\usage{
single_impute_iters(
  x,
  y,
  y_miss,
  y_real = NULL,
  imp_model = "glmnet_pred",
  lo = -Inf,
  hi = Inf,
  iters_each = 100,
  gibbs = c()
)
}
\arguments{
\item{x}{dataset}

\item{y}{dataset}

\item{y_miss}{y including missing data}

\item{y_real}{real y data}

\item{imp_model}{imputation model}

\item{lo}{lower end of truncated distribution}

\item{hi}{upper end of truncated distribution}

\item{iters_each}{number of iterations}

\item{gibbs}{gibbs sampling method}
}
\description{
Single missing variable imputation based on Gibbs sampler
}
