% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targeted_data}
\alias{targeted_data}
\title{Targeted data}
\format{
An object of class \code{data.frame} with 40 rows and 42 columns.
}
\usage{
targeted_data
}
\description{
A dataset containing targeted analysis data
}
\keyword{datasets}
