% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{Ttest_cor_cal_plot}
\alias{Ttest_cor_cal_plot}
\title{Univariate analysis and plot}
\usage{
Ttest_cor_cal_plot(
  impute_results,
  group = NULL,
  plot = T,
  x = "Miss_Prop",
  cor = "P",
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{imputation results}

\item{group}{group}

\item{plot}{plot or not}

\item{x}{x-axis label}

\item{cor}{cor}

\item{colors}{colours}

\item{shapes}{shapes}
}
\description{
Univariate analysis and plot
}
