% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{Ttest_P_cal_plot}
\alias{Ttest_P_cal_plot}
\title{Ttest_P_cal_plot}
\usage{
Ttest_P_cal_plot(
  impute_results,
  group = NULL,
  plot = T,
  p_cut = 0.05,
  x = "Miss_Prop",
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{imputation results}

\item{group}{group}

\item{plot}{plot or not}

\item{p_cut}{p_cut}

\item{x}{x-axis label}

\item{colors}{colours}

\item{shapes}{shapes}
}
\description{
Ttest_P_cal_plot
}
