% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{untargeted_data}
\alias{untargeted_data}
\title{Untargeted data}
\format{
An object of class \code{data.frame} with 37 rows and 113 columns.
}
\usage{
untargeted_data
}
\description{
A dataset containing untargeted analysis data
}
\keyword{datasets}
